refactor(python/examples): split flask example into app/config/middleware
Renames examples/simple-server/ to examples/flask/, extracts the env-driven
Config into config.py, lifts the @mpp_charge decorator into middleware.py,
and wires app.py as a create_app(mpp, settings) factory. Drops the Sinatra
mirror callouts now that the layout follows Flask best practices.
